Separate out Story Card View
