+++ src/scss/components/cards/_story-card.scss
@@ -80,7 +80,7 @@
       }
 
       .ood-story-card__image {
-        transform: scale($ood-image-zoom);
+        transform: scale($ood-card-image-zoom);
       }
     }
   }
