+++ src/scss/utilities/variables/_cards.scss
@@ -3,7 +3,7 @@
 // All variables used in cards
 
 // Card hover/focus image zoom effect scale
-$ood-image-zoom: 1.03;
+$ood-card-image-zoom: 1.03;
 
 // Story Card
 $ood-horizontal-story-card-image-width: 40vw;
